Add files via upload
Instagram Downloader Project - .exe File
This is a simple Instagram downloader application created using Python. The program allows users to download profile from Instagram by entering the Username , you can view the Instagram profile .
Attention:
This application may not work properly in regions where Instagram is blocked or restricted due to network issues.
The program is designed to be simple and functional, focusing on basic features for downloading Instagram content.
It is a beginner-level project that demonstrates how to use Python for building small applications.
Limitations:
As Instagram may be blocked in some regions (such as in Iran), the functionality of the app might be affected.
This was my first project from 1.5 years ago , I just wanted to mention , it was the starting point for my journey in coding and even though it is simple, it means a lot to me.